Optimized replay file size and minor code clean up

+++ src/main/java/me/jumper251/replay/replaysystem/data/types/AnimationData.java
@@ -7,13 +7,13 @@ public class AnimationData extends PacketData {
 	 */
 	private static final long serialVersionUID = -5227638148471461255L;
 
-	private int id;
+	private byte id;
 	
 	public AnimationData(int id) {
-		this.id = id;
+		this.id = (byte) id;
 	}
 	
-	public int getId() {
+	public byte getId() {
 		return id;
 	}
 }
