+++ src/main/java/me/jumper251/replay/replaysystem/data/types/MovingData.java
@@ -2,21 +2,20 @@ package me.jumper251.replay.replaysystem.data.types;
 
 
 public class MovingData extends PacketData {
-
 	
 	/**
 	 * 
 	 */
 	private static final long serialVersionUID = -3792160902735306458L;
 
-	private double x, y, z;
+	private float x, y, z;
 	
 	private float pitch, yaw;
 	
 	public MovingData(double x, double y, double z, float pitch, float yaw) {
-		this.x = x;
-		this.y = y;
-		this.z = z;
+		this.x = (float) x;
+		this.y = (float) y;
+		this.z = (float) z;
 		this.pitch = pitch;
 		this.yaw = yaw;
 	}
@@ -40,7 +39,5 @@ public class MovingData extends PacketData {
 	public double getZ() {
 		return z;
 	}
-
-	
 	
 }
